BB-40 refactor dir
