--- src/test/java/com/butbetter/applicationservices/ApplicationServicesApplicationTests.java
@@ -1,13 +1,0 @@
-package com.butbetter.applicationservices;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationServicesApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
